Move addNewField() to useField() hook
This probably makes more sense
than in the useBlock() hook.
